Update postprocessors.py
increase mesh output faces to 150k instead of 40k for finer details. 

Increases runtime massively, but the finished meshs output by it are vastly superior. 

I can add in a slider or dropdown to teh main UX if you want to integrate some user choice and suggest that as a parrallel change if you let me know that'd be handy?